refactor(composition): split monolithic register-events into cohesive wiring modules
