+++ src/composition/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Composition module barrel.
+ * Exposes the public API used by the plugin entry point.
+ */
+export * from "./container";
+export * from "./register-commands";
+export * from "./register-events";
+export * from "./register-settings";
+export * from "./register-styles";
