+++ src/composition/register-commands.ts
@@ -3,12 +3,15 @@ import { registerAgileDashboardView } from "@features/agile-dashboard-view";
 import { registerTemplatingDynamicCommands } from "@features/templating-engine";
 import { registerTaskAssignmentDynamicCommands } from "@features/task-assignment";
 
+/**
+ * Registers all user-facing commands.
+ * Safe to call early; will opportunistically inject TaskIndex/OrgStructure ports if available.
+ */
 export async function registerAllCommands(container: Container) {
 	// Feature: Agile dashboard view
 	await registerAgileDashboardView(container);
 
 	// Feature: Dynamic template commands based on cursor context
-	// IMPORTANT: Wire task index into templating so insertion workflows can resolve blockRef -> TaskItem
 	await registerTemplatingDynamicCommands(
 		container.app,
 		container.plugin as any,
